Refactor hotspot/os_perf for BSD and macOS
Split implementation into separate source files for macOS and BSD where
necessary, and consolidate the code that's common between them into a
common source.

- Split CPUPerformanceInterface impl for BSD and macOS, but keep the
  common PIMPL wrapper class.
- Move SystemProcessInterface::SystemProcess to common code, implement
  functions to imitate relevant macOS proc.h function for the BSD's.
- Drop retreiving process args for FreeBSD and NetBSD for now, this was
  not implemented for macOS or OpenBSD in any case. Can add it back
  later if it has value.
- Move NetworkPerformanceInterface::NetworkPerformance to common code.
  Only minor adjustments via conditional sections between macOS and BSD
  implementation.

This work is sponsored by The FreeBSD Foundation
